add sendImmediately to true
I have no idea why this works, because per request documentation sendImmediately with auth is defaulted to true, however, on request 2.51.0, and node 0.10.35, without the sendImmediately it doesnt work.  The only downside is an extra line of code that probably sets a default value to the same option, but it seems to work on some boxes.